Move Redeem All button into the Redemptions tab of the DAO
